requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
arabic-reshaper==3.0.0
asgiref==3.11.1
asn1crypto==1.5.1
attrs==25.4.0
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
cssselect2==0.9.0
diff-match-patch==20241021
dj-database-url==3.1.2
Django==5.2.11
django-appconf==1.2.0
django-axes==8.3.1
django-import-export==4.4.0
django_compressor==4.6.0
et_xmlfile==2.0.0
fonttools==4.61.1
freetype-py==2.5.1
frozenlist==1.8.0
gunicorn==25.1.0
html5lib==1.1
idna==3.11
lxml==6.0.2
multidict==6.7.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==26.0
phonenumbers==9.0.25
pillow==12.1.1
propcache==0.4.1
psycopg2-binary==2.9.11
pycairo==1.29.0
pycparser==3.0
pydyf==0.12.1
pyHanko==0.34.1
pyhanko-certvalidator==0.30.1
PyJWT==2.11.0
pypdf==6.8.0
pyphen==0.17.2
python-bidi==0.4.2
python-dotenv==1.2.2
PyYAML==6.0.3
rcssmin==1.2.2
reportlab==3.6.13
requests==2.32.5
rjsmin==1.2.5
rlPyCairo==0.4.0
six==1.17.0
stripe==14.4.1
sqlparse==0.5.5
tablib==3.9.0
tinycss2==1.5.1
tinyhtml5==2.1.0
twilio==9.10.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
webencodings==0.5.1
whitenoise==6.12.0
xhtml2pdf==0.2.11
yarl==1.23.0
zopfli==0.4.1
